pyproject.toml
[project]
name = "mustage-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.19.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
]
